+++ sws/src/mypage/components/editProfile/ChoiceContainer.jsx
@@ -15,8 +15,9 @@ export default function ChoiceContainer({ onClose, onSelectAlbum, onSelectCamera
 }
 
 const Wrapper = styled.div`
-  position: fixed;
+  position: absolute;
   top: 0;
+  left: 0;
   width: 24.375rem;
   height: 100vh;
   background: rgba(0, 0, 0, 0.3);
@@ -29,9 +30,10 @@ const Wrapper = styled.div`
 const Container = styled.div`
   display: flex;
   width: 24.375rem;
+  max-width: 24.375rem;
   padding: 2.375rem 2.4375rem;
   flex-direction: column;
-  justify-content: flex-end;
+  justify-content: center;
   align-items: center;
   gap: 0.25rem;
   border-radius: 1.25rem 1.25rem 0 0;
